Se carga la ultima versión funcional 10/10

+++ src/main/java/co/edu/uniquindio/model/Destinos.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 public class Destinos implements Serializable {
     private String nombre, ciudad, descripcion, clima;
     private ArrayList<String> imagenes = new ArrayList<>();
-    private ArrayList<Integer> calificaciones = new ArrayList<>();
+    private ArrayList<Float> calificaciones = new ArrayList<>();
     private int contReservas;
     private int contBusquedas;
     private int numero;
@@ -28,11 +28,11 @@ public class Destinos implements Serializable {
         imagenes = imagen;
     }
 
-    public void añadirCalificacion(int calificacion)
+    public void añadirCalificacion(float calificacion)
     {
         calificaciones.add(calificacion);
     }
-    public ArrayList<Integer> calificaciones(){
+    public ArrayList<Float> calificaciones(){
         return this.calificaciones;
     }
